Move components out of utils and fix searching when separateSelected is true
