+++ src/js/components/SelectMultiple/stories/Disabled.js
@@ -29,7 +29,7 @@ export const Disabled = () => {
     <Box fill align="center" pad="large" gap="large">
       <Text>SelectMultiple Disabled</Text>
       <SelectMultiple
-        sort
+        separateSelected
         width="medium"
         showSelectedInline
         dropProps={{
