+++ src/js/components/SelectMultiple/stories/ShowSelectedInline.js
@@ -25,7 +25,7 @@ export const ShowSelectedInline = () => {
     <Box fill align="center" pad="large" gap="large">
       <Text>SelectMultiple showSelectedInline</Text>
       <SelectMultiple
-        sort
+        separateSelected
         showSelectedInline
         placeholder="Select"
         options={options}
